# docs/api/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Bastion",
    "description": "ERP para pyme española. Contrato de la versión 1 de la API, bajo /api/v1/{modulo}/{recurso}. Se genera al compilar desde los controladores y los DTO: no se escribe a mano, y el cliente del frontal se genera de él.",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/organizacion/divisas/{id}": {
      "get": {
        "tags": [
          "Divisas"
        ],
        "summary": "Devuelve una divisa.",
        "operationId": "Divisas_Obtener",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identificador de la divisa.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DivisaDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DivisaDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DivisaDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          { }
        ]
      },
      "put": {
        "tags": [
          "Divisas"
        ],
        "summary": "Cambia el nombre de una divisa.",
        "operationId": "Divisas_Modificar",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identificador de la divisa.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModificarDivisaDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ModificarDivisaDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ModificarDivisaDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DivisaDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DivisaDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DivisaDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "412": {
            "description": "Precondition Failed",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "428": {
            "description": "Precondition Required",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          { }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DivisaDto": {
        "required": [
          "id",
          "codigo",
          "nombre",
          "decimales"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Identificador de la divisa.",
            "format": "uuid"
          },
          "codigo": {
            "type": "string",
            "description": "Código ISO 4217 de tres letras, en mayúsculas."
          },
          "nombre": {
            "type": "string",
            "description": "Nombre con el que se la conoce."
          },
          "decimales": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "A cuántos decimales se redondea un importe en esta divisa.",
            "format": "int32"
          }
        },
        "description": "Una divisa que esta instalación usa, tal como sale de la API."
      },
...
      "ModificarDivisaDto": {
        "required": [
          "nombre"
        ],
        "type": "object",
        "properties": {
          "nombre": {
            "maxLength": 60,
            "minLength": 0,
            "type": "string",
            "description": "Nombre con el que se la conoce."
          }
        },
        "description": "Lo que se puede cambiar de una divisa."
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "testigoDeAcceso": {
        "type": "http",
        "description": "Testigo de acceso de quince minutos, en la cabecera Authorization. Lo emite POST /api/v1/identidad/sesiones y lo renueva POST .../sesiones/renovacion, que no lo lee de aquí sino de una cookie HttpOnly. El testigo de refresco no aparece en este contrato a propósito: el navegador no tiene que poder leerlo.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}